NameSchemaService refactored
