+++ src/contracts/Event/NameSchema/AbstractContentAwareNameSchemaEvent.php
@@ -1,0 +1,29 @@
+<?php
+
+/**
+ * @copyright Copyright (C) Ibexa AS. All rights reserved.
+ * @license For full copyright and license information view LICENSE file distributed with this source code.
+ */
+declare(strict_types=1);
+
+namespace Ibexa\Contracts\Core\Event\NameSchema;
+
+use Ibexa\Contracts\Core\Repository\Values\Content\Content;
+
+abstract class AbstractContentAwareNameSchemaEvent extends AbstractNameSchemaEvent
+{
+    protected Content $content;
+
+    public function __construct(
+        array $schemaIdentifiers,
+        Content $content
+    ) {
+        parent::__construct($schemaIdentifiers);
+        $this->content = $content;
+    }
+
+    public function getContent(): Content
+    {
+        return $this->content;
+    }
+}
